refactor(cypress): extract graphql mock commands into test utils
