+++ cypress/integration/template-starter/main.js
@@ -2,10 +2,7 @@ import { URL_EXAMPLES_STARTER } from '../../support/urls';
 
 describe('Main view', () => {
   it('should render page', () => {
-    cy.login();
-    cy.get('[data-test=top-navigation').should('exist');
-    cy.visit(`${Cypress.config('baseUrl')}${URL_EXAMPLES_STARTER}`);
-    cy.url().should('include', URL_EXAMPLES_STARTER);
+    cy.login({ redirectToUri: URL_EXAMPLES_STARTER });
     cy.findByText('Hello, world').should('exist');
     cy.percySnapshot();
   });
